Dev to Main

+++ SP1/src/main/java/app/dto/external/MovieInfoTMDBDTO.java
@@ -15,29 +15,21 @@ import lombok.NoArgsConstructor;
 public class MovieInfoTMDBDTO {
 
     // Attributes
-    private Long id;
+    private Integer id;
     private String title;
-
     @JsonProperty("original_title")
     private String originalTitle;
-
     private String overview;
-
     @JsonProperty("poster_path")
     private String posterPath;
-
     @JsonProperty("backdrop_path")
     private String backdropPath;
-
     @JsonProperty("original_language")
     private String originalLanguage;
-
     @JsonProperty("release_date")
     private String releaseDate;
-
     private boolean adult;
     private Integer budget;
-
     @JsonProperty("runtime")
     private Integer runTime;
