+++ SP1/src/main/java/app/entity/Crew.java
@@ -1,6 +1,5 @@
 package app.entity;
 
-import app.enums.CreditTitleEnum;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,21 +14,23 @@ import lombok.NoArgsConstructor;
 @Table(name="crews")
 public class Crew {
 
+    // Crew.. Anything but the actor.
+
     // Many (Crew) One (Movie)
     // Not using enum anymore as department does the same from JSON.
     // Can be used to sort later if needed (enum).
 
+    // 1 -> Female | 2 -> Male
+
     @Id
     @Column(name = "id", unique = true, nullable = false)
-    private Long id;
-    private Integer gender;                       // 1 -> Female | 2 -> Male
+    private Integer id;
+    private Integer gender;
     private String department;
     private String job;
     private String name;
-
     @Column(name = "credit_id")
     private String creditId;
-
     @ManyToOne
     @JoinColumn(name = "movie_id")
     private Movie movie;
