+++ SP1/src/main/java/app/service/external/CreditsTMDBService.java
@@ -18,8 +18,12 @@ public class CreditsTMDBService {
     private static final String BASE_LANGUAGE = "&language=en-US";
 
     // __________________________________________________________
+    // Functionality of Backend (3)
+    // ___________________________
+    //
+    // Fetch Crew & Cast from A specific movieId and store in DB
 
-    public CompletableFuture<MovieCreditsTMDBDTO> getCreditsByMovieId(Long movieId) {
+    public CompletableFuture<MovieCreditsTMDBDTO> getCreditsByMovieId(Integer movieId) {
 
         // Pre-validation
         if (movieId == null) {
